+++ scripts/upgrade-provisioner.py
@@ -186,6 +186,69 @@ def install_lb_controller(cluster_name, account_id, dry_run):
                 " --repo https://aws.github.io/eks-charts")
     execute_command(command, dry_run)
 
+def enable_minsize_zero(cluster_name, provider, dry_run):
+    print("[INFO] Enabled min_size 0 value:", end=" ", flush=True)
+    if dry_run:
+        print("DRY-RUN")
+    else:
+        rbacRoleCA = {
+            "apiVersion": "rbac.authorization.k8s.io/v1",
+            "kind": "Role",
+            "metadata": {
+                "labels": {
+                    "app.kubernetes.io/name": "clusterapi-cluster-autoscaler"
+                },
+                "name": "cluster-autoscaler-clusterapi-cluster-autoscaler",
+                "namespace": "cluster-" + cluster_name
+                },
+                "rules": [
+                {
+                    "apiGroups": [
+                        "infrastructure.cluster.x-k8s.io"
+                    ],
+                    "resources": [
+                        provider + "machinetemplates"
+                    ],
+                    "verbs": [
+                        "get",
+                        "list",
+                        "watch"
+                    ]
+                }
+            ]
+        }
+        rbacRoleBindingCA = {
+            "apiVersion": "rbac.authorization.k8s.io/v1",
+            "kind": "RoleBinding",
+            "metadata": {
+                "labels": {
+                    "app.kubernetes.io/name": "clusterapi-cluster-autoscaler"
+                },
+                "name": "cluster-autoscaler-clusterapi-cluster-autoscaler",
+                "namespace": "cluster-" + cluster_name
+            },
+            "roleRef": {
+                "apiGroup": "rbac.authorization.k8s.io",
+                "kind": "Role",
+                "name": "cluster-autoscaler-clusterapi-cluster-autoscaler"
+            },
+            "subjects": [
+                {
+                    "kind": "ServiceAccount",
+                    "name": "cluster-autoscaler-clusterapi-cluster-autoscaler",
+                    "namespace": "kube-system"
+                }
+            ]
+        }
+        with open('./rbacCA.yaml', 'a') as rbacCA_file:
+            rbacCA_file.write('\n---\n')
+            yaml.dump(rbacRoleCA, rbacCA_file, default_flow_style=False)
+            rbacCA_file.write('\n---\n')
+            yaml.dump(rbacRoleBindingCA, rbacCA_file, default_flow_style=False)
+        command = "kubectl apply -f rbacCA.yaml"
+        execute_command(command, dry_run)
+        os.remove('./rbacCA.yaml')
+
 def upgrade_capx(kubeconfig, managed, provider, namespace, version, env_vars, dry_run):
     print("[INFO] Upgrading " + namespace.split("-")[0] + " to " + version + " and capi to " + CAPI + ":", end =" ", flush=True)
     # Check if capx & capi are already upgraded
@@ -263,7 +326,7 @@ def cluster_operator(kubeconfig, helm_repo, provider, credentials, cluster_name,
                 print("FAILED")
                 print("[ERROR] " + errors)
 
-        print("[INFO] Applying new ClusterConfig CRD", end =" ", flush=True)
+        print("[INFO] Applying new ClusterConfig CRD:", end =" ", flush=True)
         # https://github.com/Stratio/cluster-operator/blob/master/deploy/helm/cluster-operator/crds/installer.stratio.com_clusterconfigs.yaml
         clusterconfig_crd = "eJzFVkuP0zAQvvdXjMRhLyTdwgXlhgqHFQ+hLdpr5CST1NSxgx9lC+K/M7bTR9J0VYFW+FTPjL+Z+eaRJkkyYx1/QG24khnQb3y0KP3NpJs3JuVqvl3MNlxWGSydsaq9R6OcLvEd1lxyS5azFi2rmGXZDIBJqSzzYuOvAKWSVishUCcNynTjCiwcFxXqAL53vb1NF6/SBT2RrMUMSkHeUNPrmjcUhzSWeZDUWO3h01K1M9Nh6b00Wrkug2mjiNhH02cSwZcBPMgFN/bDue4jiYO+E04zMQ4rqAyXjRNMj5SkM6XqKJXP3n3HSqxI1iccwkn6XLeLghhcRLRyjS2LwQLQe/n2y93D69VADFChKTXvbOBuEDJwA3aNEF9ArXS4DgMHAj1gdZrcaMv3HMVz0hcn0pHnGx9ctCIFNQRG532WWPX5gKpJTpFp7DQalLFFBsDgjZgEVXzD0qawQu1hwKyVE5XvI7paQihVI/nPAzZ5VMGpYBb7ch0Pl5S1ZAK2TDh8SQ4qaNmOYLwXcPIEL5iYFD4pjfSwVhmsre1MNp833O7ngXqqddT5u3lobV44q7SZV7hFMTe8SZgu19wSutM4JxqTELoMM5G21QvdT5C5GcRqd75ZqHOpn04UoWOfqIDvWl9z1j+NWRyJ9iLPzv371VfYuw7FGLMfeD8+NMcSeMKID9SxiLVWbcBEWXWKGO47jNOrEahxRcutr/t3otb6WqWwDEsCCgTX0d7AKoU7SdIWxZIZfPYCeKZN4om9rgSn+21sHFk7Uex30oV6DWZ1RbaDuSFTrn1n03ygn4fzTbU/01PrD25MLor8uHjHBj6kmjlhM6iZMHimjqkVSglk4yHtF0nuvTMiPuctazDfTu2KJzidhPorkJhmJ5jEPK638/eXyfKnZY+5k2tkwq53UwbBhLeuzWBxezttQB/DYDCtjgn4ZdSgvpDeWSvFyPmWeiHX2NC3SE+E9w+1/KH0hih/PtoOsV2k7f/w6teRH7Rh1MnFjrxq8Glmnbl69IP1YPhVYfy+vWr6J2M4E0a8DKx2sTHoD5ymaT2VuOLwPdrH3mcCv37P/gDZnAI8"
         clusterconfig_decoded = zlib.decompress(base64.b64decode(clusterconfig_crd))
@@ -550,8 +613,14 @@ if __name__ == '__main__':
         else:
             print("[WARN] AWS LoadBalancer Controller is only supported for EKS managed clusters")
             sys.exit(0)
-    if not config["yes"]:
-        request_confirmation()
+        if not config["yes"]:
+            request_confirmation()
+
+    # Enable min_size = 0
+    if not (provider == "azure" and managed):
+        enable_minsize_zero(cluster_name, provider, config["dry_run"])
+        if not config["yes"]:
+            request_confirmation()
 
     # CAPX
     upgrade_capx(kubeconfig, managed, provider, namespace, version, env_vars, config["dry_run"])
@@ -568,3 +637,4 @@ if __name__ == '__main__':
         restore_capsule(config["dry_run"])
 
     print("[INFO] Upgrade process finished successfully")
+